+++ lib/food_info/errors.rb
@@ -2,4 +2,5 @@ module FoodInfo
   class NoAdapterSpecified < StandardError; end
   class UnsupportedAdapter < StandardError; end
   class AuthorizationError < StandardError; end
+  class DataSourceException < StandardError; end
 end