+++ HISTORY.markdown
@@ -1,4 +1,7 @@
 # Changelog
 
+## unreleased
+* Added propagation of errors from data source
+
 ## 0.0.1  September 10, 2011
 * Initial release. FatSecret adapter supports both <tt>search</tt> and <tt>details</tt> methods.